Move packages: tftp -> itftp, http -> ihttp:
The renaming of the tftp and http packages is being made
in accordance with this guidance from the Go team,

"avoid using the same name as popular standard packages like io or http."
-- https://go.dev/blog/package-names
