+++ itftp/itftp.go
@@ -1,5 +1,5 @@
 // Package tftp implements a TFTP server for iPXE binaries.
-package tftp
+package itftp
 
 import (
 	"bytes"
